+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Added String.serializeAttributes(obj)
 * Added String#encodeHTML and String#decodeHTML
 * Added Object.flatten(obj)
+* Added Array#unique()
 
 ## 0.1.1 (2014-06-23)